move react hook into utils module
